+++ src/main.js
@@ -2,6 +2,5 @@
 
 require('./check');
 require('./upload-stat.min');
-require('./resizer');
 require('./upload');
-require('./pictures');
+require('./pictures/pictures');
